+++ src/Core/VRChatContentPublisher.Core/Services/PublishTask/ContentPublisher/AvatarContentPublisher.cs
@@ -1,5 +1,7 @@
-﻿using Microsoft.Extensions.Logging;
+﻿using MessagePipe;
+using Microsoft.Extensions.Logging;
 using VRChatContentPublisher.ConnectCore.Services;
+using VRChatContentPublisher.Core.Events.UserSession;
 using VRChatContentPublisher.Core.Models;
 using VRChatContentPublisher.Core.Models.VRChatApi;
 using VRChatContentPublisher.Core.Models.VRChatApi.Rest.Avatars;
@@ -17,14 +19,20 @@ public sealed class AvatarContentPublisher(
     string unityVersion,
     UserSessionService userSessionService,
     ILogger<AvatarContentPublisher> logger,
-    IFileService tempFileService)
-    : IContentPublisher
+    IFileService tempFileService,
+    ISubscriber<SessionStateChangedEvent> sessionStateChangedSubscriber
+) : IContentPublisher
 {
     private readonly VRChatApiClient _apiClient = userSessionService.GetApiClient();
 
     public string GetContentType() => "avatar";
     public string GetContentName() => name;
     public string GetContentPlatform() => platform;
+
+    public bool CanPublish()
+    {
+        return userSessionService.State == UserSessionState.LoggedIn;
+    }
 
     public ValueTask BeforePublishTaskAsync(string? thumbnailFileId,
         string? description,
@@ -50,6 +58,14 @@ public sealed class AvatarContentPublisher(
         PublishStageProgressReporter? progressReporter = null,
         CancellationToken cancellationToken = default)
     {
+        using var sessionValidScope = new EnsureSessionValidScope(
+            userSessionService.UserNameOrEmail,
+            sessionStateChangedSubscriber,
+            cancellationToken
+        );
+
+        cancellationToken = sessionValidScope.CancellationToken;
+
         await using var bundleFileStream = await tempFileService.GetFileAsync(bundleFileId);
         var thumbnailFile = thumbnailFileId is not null
             ? await tempFileService.GetFileWithNameAsync(thumbnailFileId)
@@ -197,7 +213,11 @@ public sealed class AvatarContentPublisher(
     }
 }
 
-public sealed class AvatarContentPublisherFactory(ILogger<AvatarContentPublisher> logger, IFileService tempFileService)
+public sealed class AvatarContentPublisherFactory(
+    ILogger<AvatarContentPublisher> logger,
+    IFileService tempFileService,
+    ISubscriber<SessionStateChangedEvent> sessionStateChangedSubscriber
+)
 {
     public AvatarContentPublisher Create(
         UserSessionService userSession,
@@ -213,7 +233,8 @@ public sealed class AvatarContentPublisherFactory(ILogger<AvatarContentPublisher
             unityVersion,
             userSession,
             logger,
-            tempFileService
+            tempFileService,
+            sessionStateChangedSubscriber
         );
     }
 }